Show classes found

+++ ThePersonalTrainerClub/app/src/main/java/es/exceptioncoders/thepersonaltrainerclub/model/model/ClassModel.kt
@@ -1,12 +1,12 @@
 package es.exceptioncoders.thepersonaltrainerclub.model.model
 
-data class ClassModel (
+data class ClassModel(
         val id: String,
         val instructor: TrainerModel,
         val sport: SportModel,
         val location: LocationModel,
         val description: String,
-        val price: Float,
+        val price: Double,
         val maxusers: Int,
         val duration: Int,
         val registered: Int?,
