+++ ThePersonalTrainerClub/app/src/main/java/es/exceptioncoders/thepersonaltrainerclub/model/model/LocationModel.kt
@@ -3,6 +3,6 @@ package es.exceptioncoders.thepersonaltrainerclub.model.model
 data class LocationModel(
         val id: String?,
         val type: String,
-        val coordinates: Array<Float>,
+        val coordinates: Array<Double>,
         val description: String?
 )
